feat: 각 서버와의 연동 기능 추가

+++ src/backend/user-service/Business/Auth/service/AuthService.cs
@@ -1,14 +1,12 @@
-using System.Text;
 using Castle.DynamicProxy;
 using MimeKit;
-using Newtonsoft.Json;
 using user_service.auth.dto;
 using user_service.auth.repository;
 using user_service.Business.Auth.service;
 using user_service.common;
 using user_service.common.exception;
 using user_service.intercepter;
-using user_service.user.dto;
+using user_service.common.dto;
 
 namespace user_service.auth.service;
 
@@ -17,72 +15,49 @@ public class AuthService : IAuthService
     private readonly IUserRepository _userRepository;
     private IAuthRedisRepository _redis = null!;
     private IConfiguration _config = null!;
-    private HttpClient _communityClient = null!;
+    private ICommunityClient _communityClient = null!;
     public AuthService(IUserRepository userRepository,
                     IAuthRedisRepository redisRepository,
                     IConfiguration config,
+                    ICommunityClient communityClient,
                     LogInterceptor interceptor)
     {
         var generator = new ProxyGenerator();
 
         _userRepository = generator.CreateInterfaceProxyWithTargetInterface<IUserRepository>(userRepository, interceptor);
+        _communityClient = generator.CreateInterfaceProxyWithTargetInterface<ICommunityClient>(communityClient, interceptor);
         _redis = redisRepository;
         _config = config;
-        _communityClient = new HttpClient();
-        _communityClient.BaseAddress = new Uri("http://34.22.109.45:8000/api/community/registration/user");
     }
 
-    public async Task RegisterAsync(RegisterDTO register)
+    public async Task RegisterAsync(RegisterDTO register, string traceId)
     {
         // 체크
-        SameEmailCheck(register.Email);
-        CheckEmailChecksum(register.Email, register.EmailChecksum);
+        // SameEmailCheck(register.Email);
+        // CheckEmailChecksum(register.Email, register.EmailChecksum);
 
-        _redis.DeleteChecksum(register.Email);
+        // _redis.DeleteChecksum(register.Email);
 
         // 비밀번호 암호화
         register.Password = Utils.SHA256Hash(register.Password);
         if (_userRepository.InsertUser(register) == false)
             throw new ServiceException(4010);
-        
+
         UserModel? user = _userRepository.GetUserByEmail(register.Email);
-        if(user == null)
+        if (user == null)
             throw new ServiceException(4010);
 
-        using StringContent jsonContent = new(
-            JsonConvert.SerializeObject(new CommunityDTO(){
-                userId = user.Id,
-                email = user.Email,
-                name = user.Name,
-                profile = user.Profile
-            }),
-            Encoding.UTF8,
-            "application/json"
-        );
-
-        CancellationTokenSource cts = new CancellationTokenSource();
-        cts.CancelAfter(100); // Cancel after 1 second
-        CancellationToken cancellationToken = cts.Token;
-
-        try
+        bool isSuccess = await _communityClient.RegisterUserAsync(new CommunityUserDTO()
         {
-            HttpResponseMessage response = await _communityClient.PostAsync("", jsonContent, cancellationToken);
-            string str = await response.Content.ReadAsStringAsync();
-            var dto = JsonConvert.DeserializeObject<CommunityResponseDTO<CommunityBaseException>>(str);
+            userId = user.Id,
+            email = user.Email,
+            name = user.Name,
+            profile = user.Profile
+        }, traceId);
 
-            if(dto!.ResultCode != 200)
-            {
-                System.Console.WriteLine(dto.ResultData!.Exception);
-                // var error = JsonConvert.DeserializeObject<CommunityBaseException>(dto.ResultData);
-                // System.Console.WriteLine(error!.Exception);
-                throw new ServiceException(4010);
-            }
-        }
-        catch (Exception e)
+        if(isSuccess == false)
         {
-            System.Console.WriteLine(e.Message);
             _userRepository.DeleteUser(user.Id);
-
             throw new ServiceException(4010);
         }
     }
